Доработки по замечаниям наставника
